Fix a deadlock in the breakpoint code.
In froyo we started using "hard" breakpoints, where we replace the
existing opcodes with breakpoint instructions.  This requires some
coordination to avoid confusing the verifier.  The previous approach
allowed the breakpoints to be inserted, and "undid" them while the
verifier ran; this worked, but caused us to be holding a lock for
an extended period.

The new approach just avoids altering the bytecode of unverified
classes, and then "flushes" the breakpoint set out between the time
when verification completes and class initialization starts.  This
removes the possibility of blocking with the lock held, and makes
everything much simpler.

For bug 2615063.

Change-Id: I7f43e09a755fba27b335454659b3f04e8b2179ac