Update sqlite java 6 driver to coincide with changes made upstream.
Some of these changes are kind of arbitrary, but where it didn't matter I went with the upstream
version to make future merges easier.

Changes NOT made upstream are still marked with android-changed, but I updated the rest of them.

Tried to fix whitespace as much as possible. Tell me if I missed any.

Change-Id: I3396bd26d2cc00306b3867cea49d100430691bd1