Remove File's cached UTF-8 byte[].
This doesn't really improve File.getCanonicalPath and its helpers much, but it
improves everything else, and feels like a step in the right direction.

Benchmarking shows only a tiny difference in the uncommon case of a reused
File, and a significant increase in performance in the more likely case of
a single-use File.

Before:

      benchmark    us
         Exists 26.14
    ExistsReuse  5.08

After:

      benchmark    us
         Exists 13.90
    ExistsReuse  5.23

Bug: 2281992
Change-Id: Ibc6fd2581983ae8a57bf4b7f49196c76ad41169e